Module jf_plonk::circuit

Interfaces for Plonk-based constraint systems

Re-exports

pub use basic::PlonkCircuit;

Modules

basic

Basic instantiations of Plonk-based constraint systems

customized

Customized gates and gadgets for rescue hash related, elliptic curve related, rescue-based transcript and lookup table etc.

gates

Implementations of various Gates.

Traits

Arithmetization

An interface that transforms Plonk circuits to polynomial used by Plonk-based SNARKs.

Circuit

An interface for Plonk constraint systems.

Type Definitions

GateId

An index to a gate in circuit.

Variable

An index to one of the witness values.

WireId

An index to the type of gate wires. There are 4 different types of input gate wires (with indices 0..3), 1 type of output gate wires (with index 4), and 1 type of lookup gate wires (with index 5).