Struct jf_plonk::circuit::basic::PlonkCircuit

pub struct PlonkCircuit<F> where
    F: FftField,  { /* private fields */ }

A specific Plonk circuit instantiation.

Implementations

impl<F: FftField> PlonkCircuit<F>

pub fn new_turbo_plonk() -> Self

Construct a new TurboPlonk circuit.

pub fn new_ultra_plonk(range_bit_len: usize) -> Self

Construct a new UltraPlonk circuit.

pub fn insert_gate(
&mut self,
wire_vars: &[Variable; 5],
gate: Box<dyn Gate<F>>
) -> Result<(), PlonkError>

Insert a general (algebraic) gate

wire_vars - wire variables. Each of these variables must be in range

gate - specific gate to be inserted

returns - an error if some verification fails

pub fn check_var_bound(&self, var: Variable) -> Result<(), PlonkError>

Checks if a variable is strictly less than the number of variables. This function must be invoked for each gate as this check is not applied in the function insert_gate

var - variable to check

returns - Error if the variable is out of bound (i.e. >= number of variables)

pub fn check_vars_bound(&self, vars: &[Variable]) -> Result<(), PlonkError>

Check if a list of variables are strictly less than the number of variables.

vars - variables to check

returns - Error if the variable is out of bound (i.e. >= number of variables)

pub fn witness_mut(&mut self, idx: Variable) -> &mut F

Change the value of a variable. Only used for testing.

pub fn range_bit_len(&self) -> Result<usize, PlonkError>

The bit length of UltraPlonk range gates.

pub fn range_size(&self) -> Result<usize, PlonkError>

The range size of UltraPlonk range gates.

impl<F: PrimeField> PlonkCircuit<F>

Methods for finalizing and merging the circuits.

pub fn finalize_for_arithmetization(&mut self) -> Result<(), PlonkError>

Finalize the setup of the circuit before arithmetization.

pub fn finalize_for_mergeable_circuit(
&mut self,
circuit_type: MergeableCircuitType
) -> Result<(), PlonkError>

Finalize the setup of a mergeable circuit. Two circuits can be merged only if they are with different circuit types The method only supports TurboPlonk circuits.

impl<F> PlonkCircuit<F> where
F: PrimeField,

pub fn glv_mul<P: TEModelParameters<BaseField = F> + Clone>(
&mut self,
scalar: Variable,
base: &PointVariable
) -> Result<PointVariable, PlonkError>

Perform GLV multiplication in circuit (which costs a few less constraints).

impl<F> PlonkCircuit<F> where
F: PrimeField,

pub fn point_witness(
&self,
point_var: &PointVariable
) -> Result<Point<F>, PlonkError>

Return the witness point for the circuit

pub fn create_point_variable(
&mut self,
point: Point<F>
) -> Result<PointVariable, PlonkError>

Add a new EC point (as witness) to the circuit

pub fn create_constant_point_variable(
&mut self,
point: Point<F>
) -> Result<PointVariable, PlonkError>

Add a new EC point (as a constant) to the circuit

pub fn create_public_point_variable(
&mut self,
point: Point<F>
) -> Result<PointVariable, PlonkError>

Add a new EC point (as public input) to the circuit

pub fn point_equal_gate(
&mut self,
point0: &PointVariable,
point1: &PointVariable
) -> Result<(), PlonkError>

Constrain two point variables to be the same. Return error if the input point variables are invalid.

pub fn check_equal_point(
&mut self,
point0: &PointVariable,
point1: &PointVariable
) -> Result<Variable, PlonkError>

Obtain a bool variable representing whether two point variables are equal. Return error if point variables are invalid.

impl<F> PlonkCircuit<F> where
F: PrimeField,

pub fn inverse_point(
&mut self,
point_var: &PointVariable
) -> Result<PointVariable, PlonkError>

Inverse a point variable

pub fn neutral_point_variable(&self) -> PointVariable

Return the point variable for the infinity point in the TE form: (0, 1)

pub fn is_neutral_point<P>(
&mut self,
point_var: &PointVariable
) -> Result<Variable, PlonkError> where
P: Parameters<BaseField = F> + Clone,

Obtain a boolean variable indicating whether a point is the neutral point (0, 1) Return variable with value 1 if it is, or 0 otherwise Return error if input variables are invalid

pub fn on_curve_gate<P: Parameters<BaseField = F> + Clone>(
&mut self,
point_var: &PointVariable
) -> Result<(), PlonkError>

Constrain a point to be on certain curve, namely its coordinates satisfy the curve equation, which is curve-dependent. Currently we only support checks of a GroupAffine::
over a base field which is the bls12-381 scalar field

Returns error if input variables are invalid

pub fn ecc_add<P: Parameters<BaseField = F> + Clone>(
&mut self,
point_a: &PointVariable,
point_b: &PointVariable
) -> Result<PointVariable, PlonkError>

Obtain a variable to the point addition result of point_a + point_b where “+” is the group operation over an elliptic curve. Currently only supports GroupAffine::
addition.

Returns error if inputs are invalid

pub fn fixed_base_scalar_mul<P: Parameters<BaseField = F> + Clone>(
&mut self,
scalar: Variable,
base: &GroupAffine<P>
) -> Result<PointVariable, PlonkError>

Obtain the fixed-based scalar multiplication result of scalar * Base Currently only supports GroupAffine::
scalar multiplication.

pub fn variable_base_scalar_mul<P: Parameters<BaseField = F> + Clone>(
&mut self,
scalar: Variable,
base: &PointVariable
) -> Result<PointVariable, PlonkError>

Obtain a variable of the result of a variable base scalar multiplication. both scalar and base are variables. Currently only supports GroupAffine::
. If the parameter is bandersnatch, we will use GLV multiplication.

pub fn variable_base_binary_scalar_mul<P: Parameters<BaseField = F> + Clone>(
&mut self,
scalar_bits_le: &[Variable],
base: &PointVariable
) -> Result<PointVariable, PlonkError>

Obtain a variable of the result of a variable base scalar multiplication. Both scalar_bits_le and base are variables, where scalar_bits_le is the little-endian form of the scalar. Currently only supports GroupAffine::
.

impl<F: PrimeField> PlonkCircuit<F>

pub fn create_table_and_lookup_variables(
&mut self,
lookup_vars: &[(Variable, Variable, Variable)],
table_vars: &[(Variable, Variable)]
) -> Result<(), PlonkError>

Create a table with keys/values [0, …, n - 1] and [table_vars[0], …, table_vars[n - 1]]; and create a list of variable tuples to be looked up: [lookup_vars[0], …, lookup_vars[m - 1]];

w.l.o.g we assume n = m as we can pad with dummy tuples when n != m

impl<F: PrimeField> PlonkCircuit<F>

pub fn mod_add(
&mut self,
x: &FpElemVar<F>,
y: &FpElemVar<F>,
p: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Modular addition gate: Given Fp elements x, y and modulus p, compute z = x + y mod p

pub fn mod_add_constant(
&mut self,
x: &FpElemVar<F>,
y: &FpElem<F>,
p: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Modular addition gate: Given input x: Fp element variable, y: Fp element, and modulus p: Fp element, use y as a constant and compute z = x + y mod p

pub fn mod_add_vec(
&mut self,
x: &[FpElemVar<F>],
p: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Modular addition gate: Given Fp elements &[x] and modulus p, compute z = \sum x mod p

pub fn mod_mul(
&mut self,
x: &FpElemVar<F>,
y: &FpElemVar<F>,
p: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Modular multiplication gate: Given Fp elements x, y and modulus p, compute z = x * y mod p.

pub fn mod_mul_constant(
&mut self,
x: &FpElemVar<F>,
y: &FpElem<F>,
p: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Modular multiplication gate: Given input x: Fp element variable, y: Fp element, and modulus p: Fp element, use y as a constant compute z = x * y mod p

pub fn mod_negate(
&mut self,
x: &FpElemVar<F>,
p: &F
) -> Result<FpElemVar<F>, PlonkError>

Negate an FpElemVar mod p where p is a public variable which is also the modulus for the FpElem element.

impl<F: PrimeField> PlonkCircuit<F>

pub fn non_native_power_11_gate<T: PrimeField>(
&mut self,
x: &FpElemVar<F>,
x_to_11: &FpElemVar<F>
) -> Result<(), PlonkError>

generate a non-native circuit for the statement x^11 = y

Input:

variable representation of x over a target field T whose order is less than F.

variable representation of x^11 over a same field

Cost: 5 mod_mul + 2 equal gate

pub fn non_native_power_11_gen<T: PrimeField>(
&mut self,
x: &FpElemVar<F>
) -> Result<FpElemVar<F>, PlonkError>

generate a non-native circuit for the statement x^11 = y

Input: variable representation of x over a target field T whose order is less than F.

Output: variable representation of x^11

Cost: 5 mod_mul

pub fn non_native_power_5_gen<T: PrimeField>(
&mut self,
x: &FpElemVar<F>
) -> Result<FpElemVar<F>, PlonkError>

generate a non-native circuit for the statement x^5 = y

Input: variable representation of x over a target field T whose order is less than F.

Output: variable representation of x^5

Cost: 3 mod_mul

pub fn non_native_linear_gen<T: PrimeField>(
&mut self,
x: &[FpElemVar<F>],
y: &[FpElem<F>],
c: &FpElem<F>
) -> Result<FpElemVar<F>, PlonkError>

Input vector x and y, and a constant c, generate a non-native circuit for the statement var_output = inner_product(x, y) + c Input: variable representation of x, y, c over a target field T whose order is less than F.

Cost: 4 mod_mul_constant + 1 mod_add_internal

impl<F> PlonkCircuit<F> where
F: PrimeField,

pub fn aggregate_verify_keys<E, P>(
&mut self,
vk_type_a_vars: &[VerifyingKeyVar<E>],
vk_type_b_vars: &[VerifyingKeyVar<E>]
) -> Result<Vec<VerifyingKeyVar<E>>, PlonkError> where
E: PairingEngine,
P: TEParam<BaseField = F> + Clone,

Aggregate verification keys

impl<F: PrimeField> PlonkCircuit<F>

pub fn num_range_blocks(&self) -> Result<usize, PlonkError>

The number of range blocks, i.e., the minimal integer such that RANGE_SIZE^NUM_RANGES >= p,

impl<F> PlonkCircuit<F> where
F: PrimeField,

pub fn quad_poly_gate(
&mut self,
wires: &[Variable; 5],
q_lc: &[F; 4],
q_mul: &[F; 2],
q_o: F,
q_c: F
) -> Result<(), PlonkError>

Arithmetic gates

Quadratic polynomial gate: q1 * a + q2 * b + q3 * c + q4 * d + q12 * a * b + q34 * c * d + q_c = q_o * e, where q1, q2, q3, q4, q12, q34, q_c, q_o are selectors; a, b, c, d are input wires; e is the output wire. Return error if variables are invalid.

pub fn gen_quad_poly(
&mut self,
wires: &[Variable; 4],
q_lc: &[F; 4],
q_mul: &[F; 2],
q_c: F
) -> Result<Variable, PlonkError>

Arithmetic gates

Quadratic polynomial gate: e = q1 * a + q2 * b + q3 * c + q4 * d + q12 * a * b + q34 * c * d + q_c, where q1, q2, q3, q4, q12, q34, q_c are selectors; a, b, c, d are input wires

Return the variable for Return error if variables are invalid.

pub fn lc_gate(
&mut self,
wires: &[Variable; 5],
coeffs: &[F; 4]
) -> Result<(), PlonkError>

Constrain a linear combination gate: q1 * a + q2 * b + q3 * c + q4 * d = y

pub fn lc(
&mut self,
wires_in: &[Variable; 4],
coeffs: &[F; 4]
) -> Result<Variable, PlonkError>

Obtain a variable representing a linear combination. Return error if variables are invalid.

pub fn mul_add_gate(
&mut self,
wires: &[Variable; 5],
q_muls: &[F; 2]
) -> Result<(), PlonkError>

Constrain a mul-addition gate: q_muls[0] * wires[0] * wires[1] + q_muls[1] * wires[2] * wires[3] = wires[4]

pub fn mul_add(
&mut self,
wires_in: &[Variable; 4],
q_muls: &[F; 2]
) -> Result<Variable, PlonkError>

Obtain a variable representing q12 * a * b + q34 * c * d, where a, b, c, d are input wires, and q12, q34 are selectors. Return error if variables are invalid.

pub fn sum(&mut self, elems: &[Variable]) -> Result<Variable, PlonkError>

Obtain a variable representing the sum of a list of variables. Return error if variables are invalid.

pub fn conditional_select(
&mut self,
b: Variable,
x_0: Variable,
x_1: Variable
) -> Result<Variable, PlonkError>

Obtain a variable that equals x_0 if b is zero, or x_1 if b is one. Return error if variables are invalid.

pub fn add_constant(
&mut self,
input_var: Variable,
elem: &F
) -> Result<Variable, PlonkError>

Obtains a variable representing an addition with a constant value Return error if the input variable is invalid

pub fn mul_constant(
&mut self,
input_var: Variable,
elem: &F
) -> Result<Variable, PlonkError>

Obtains a variable representing a multiplication with a constant value Return error if the input variable is invalid

pub fn logic_or_gate(
&mut self,
a: Variable,
b: Variable
) -> Result<(), PlonkError>

Logic gates Constrain that a is true or b is true. Return error if variables are invalid.

pub fn check_equal(
&mut self,
a: Variable,
b: Variable
) -> Result<Variable, PlonkError>

Obtain a bool variable representing whether two input variables are equal. Return error if variables are invalid.

pub fn check_is_zero(&mut self, a: Variable) -> Result<Variable, PlonkError>

Obtain a bool variable representing whether input variable is zero. Return error if the input variable is invalid.

pub fn non_zero_gate(&mut self, var: Variable) -> Result<(), PlonkError>

Constrain a variable to be non-zero. Return error if the variable is invalid.

pub fn logic_neg(&mut self, a: Variable) -> Result<Variable, PlonkError>

Assuming value represented by a is boolean, obtain a variable representing the result of a logic negation gate. Return the index of the variable. Return error if the input variable is invalid.

pub fn logic_and(
&mut self,
a: Variable,
b: Variable
) -> Result<Variable, PlonkError>

Assuming values represented by a and b are boolean, obtain a variable representing the result of a logic AND gate. Return the index of the variable. Return error if the input variables are invalid.

pub fn logic_and_all(
&mut self,
vars: &[Variable]
) -> Result<Variable, PlonkError>

Given a list of boolean variables, obtain a variable representing the result of a logic AND gate. Return the index of the variable. Return error if the input variables are invalid.

pub fn logic_or(
&mut self,
a: Variable,
b: Variable
) -> Result<Variable, PlonkError>

Assuming values represented by a and b are boolean, obtain a variable representing the result of a logic OR gate. Return the index of the variable. Return error if the input variables are invalid.

pub fn enforce_true(&mut self, a: Variable) -> Result<(), PlonkError>

Assuming values represented by a is boolean. Constrain a is true

pub fn enforce_false(&mut self, a: Variable) -> Result<(), PlonkError>

Assuming values represented by a is boolean. Constrain a is false

pub fn power_11_gen(&mut self, x: Variable) -> Result<Variable, PlonkError>

Return a variable to be the 11th power of the input variable. Cost: 3 constraints.

pub fn power_11_gate(
&mut self,
x: Variable,
x_to_11: Variable
) -> Result<(), PlonkError>

Constraint a variable to be the 11th power of another variable. Cost: 3 constraints.

pub fn truncate(
&mut self,
a: Variable,
bit_length: usize
) -> Result<Variable, PlonkError>

Obtain the truncation of the input. Constrain that the input and output values congruent modulo 2^bit_length. Return error if the input is invalid.

pub fn truncate_gate(
&mut self,
a: Variable,
b: Variable,
bit_length: usize
) -> Result<(), PlonkError>

Truncation gate. Constrain that b == a modulo 2^bit_length. Return error if the inputs are invalid; or b >= 2^bit_length.

impl<F: PrimeField> PlonkCircuit<F>

pub fn range_gate(
&mut self,
a: Variable,
bit_len: usize
) -> Result<(), PlonkError>

Constrain a variable to be within the [0, 2^bit_len) range Return error if the variable is invalid.

pub fn check_in_range(
&mut self,
a: Variable,
bit_len: usize
) -> Result<Variable, PlonkError>

Return a boolean variable indicating whether variable a is in the range [0, 2^bit_len). Return error if the variable is invalid. TODO: optimize the gate for UltraPlonk.

pub fn unpack(
&mut self,
a: Variable,
bit_len: usize
) -> Result<Vec<Variable>, PlonkError>

Obtain the bit_len-long binary representation of variable a Return a list of variables [b0, …, b_bit_len] which is the binary representation of a. Return error if the a is not the range of [0, 2^bit_len).