Trait jf_plonk::circuit::customized::ecc::SWToTEConParam

pub trait SWToTEConParam: PrimeField {
    const S: Self::BigInt;
    const NEG_ALPHA: Self::BigInt;
    const BETA: Self::BigInt;
}

This trait holds constants that are used for curve conversion from short Weierstrass form to twisted Edwards form.

Required Associated Constants

const S: Self::BigInt

Parameter S.

const NEG_ALPHA: Self::BigInt

Parameter 1/alpha.

const BETA: Self::BigInt

Parameter beta.