Trait jf_plonk::circuit::customized::rescue::RescueGadget

pub trait RescueGadget<F: RescueParameter> {
    fn rescue_permutation(
        &mut self, 
        input_var: RescueStateVar
    ) -> Result<RescueStateVar, PlonkError>;
    fn prp(
        &mut self, 
        key_var: &RescueStateVar, 
        input_var: &RescueStateVar
    ) -> Result<RescueStateVar, PlonkError>;
    fn rescue_sponge_no_padding(
        &mut self, 
        data_vars: &[Variable], 
        num_output: usize
    ) -> Result<Vec<Variable>, PlonkError>;
    fn rescue_sponge_with_padding(
        &mut self, 
        data_vars: &[Variable], 
        num_output: usize
    ) -> Result<Vec<Variable>, PlonkError>;
    fn rescue_full_state_keyed_sponge_no_padding(
        &mut self, 
        key: Variable, 
        data_vars: &[Variable]
    ) -> Result<Variable, PlonkError>;
    fn key_schedule(
        &mut self, 
        mds_states: &RescueMatrix<F>, 
        key_var: &RescueStateVar, 
        prp_instance: &PRP<F>
    ) -> Result<Vec<RescueStateVar>, PlonkError>;
    fn create_rescue_state_variable(
        &mut self, 
        state: &RescueVector<F>
    ) -> Result<RescueStateVar, PlonkError>;
    fn prp_with_round_keys(
        &mut self, 
        input_var: &RescueStateVar, 
        mds: &RescueMatrix<F>, 
        keys_vars: &[RescueStateVar]
    ) -> Result<RescueStateVar, PlonkError>;
}

Trait for rescue circuit over native field.

Required Methods

fn rescue_permutation(
&mut self,
input_var: RescueStateVar
) -> Result<RescueStateVar, PlonkError>

Given an input state st_0 and an output state st_1, ensure that st_1 = rescue_permutation(st_0) where rescue_permutation is the instance of the Rescue permutation defined by its respective constants

input_var - variables corresponding to the input state

returns - variables corresponding to the output state

fn prp(
&mut self,
key_var: &RescueStateVar,
input_var: &RescueStateVar
) -> Result<RescueStateVar, PlonkError>

Rescue based Pseudo Random Permutation (PRP)

key_var - rescue state variable corresponding to the cipher key

input_var - rescue state variable corresponding to the plaintext

returns - state variable corresponding to the cipher text

fn rescue_sponge_no_padding(
&mut self,
data_vars: &[Variable],
num_output: usize
) -> Result<Vec<Variable>, PlonkError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables, data_vars.len() should be a positive integer that is a multiple of the sponge rate (i.e. 3)

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_sponge_with_padding(
&mut self,
data_vars: &[Variable],
num_output: usize
) -> Result<Vec<Variable>, PlonkError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables,

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_full_state_keyed_sponge_no_padding(
&mut self,
key: Variable,
data_vars: &[Variable]
) -> Result<Variable, PlonkError>

Full-State-Keyed-Sponge with a single output

key - key variable

input - input variables,

returns a variable that refers to the output

fn key_schedule(
&mut self,
mds_states: &RescueMatrix<F>,
key_var: &RescueStateVar,
prp_instance: &PRP<F>
) -> Result<Vec<RescueStateVar>, PlonkError>

Return the round keys variables for the Rescue block cipher

mds_states - Rescue MDS matrix

key_var - state variable representing the cipher key

returns - state variables corresponding to the scheduled keys

fn create_rescue_state_variable(
&mut self,
state: &RescueVector<F>
) -> Result<RescueStateVar, PlonkError>

Create a variable representing a rescue state

state - Rescue state

returns - state variables corresponding to the state

fn prp_with_round_keys(
&mut self,
input_var: &RescueStateVar,
mds: &RescueMatrix<F>,
keys_vars: &[RescueStateVar]
) -> Result<RescueStateVar, PlonkError>

Return the variable corresponding to the output of the of the Rescue PRP where the rounds keys have already been computed “dynamically”

input_var - variable corresponding to the plain text

mds_states - Rescue MDS matrix

key_vars - variables corresponding to the scheduled keys

returns -