Trait jf_plonk::circuit::Circuit

pub trait Circuit<F: Field> {

    fn num_gates(&self) -> usize;

    fn num_vars(&self) -> usize;

    fn num_inputs(&self) -> usize;

    fn num_wire_types(&self) -> usize;

    fn public_input(&self) -> Result<Vec<F>, PlonkError>;

    fn check_circuit_satisfiability(
        &self, 
        pub_input: &[F]
    ) -> Result<(), PlonkError>;

    fn create_constant_variable(
        &mut self, 
        val: F
    ) -> Result<Variable, PlonkError>;

    fn create_variable(&mut self, val: F) -> Result<Variable, PlonkError>;

    fn create_public_variable(&mut self, val: F) -> Result<Variable, PlonkError>;

    fn set_variable_public(&mut self, var: Variable) -> Result<(), PlonkError>;

    fn zero(&self) -> Variable;

    fn one(&self) -> Variable;

    fn witness(&self, idx: Variable) -> Result<F, PlonkError>;

    fn constant_gate(
        &mut self, 
        var: Variable, 
        constant: F
    ) -> Result<(), PlonkError>;

    fn add_gate(
        &mut self, 
        a: Variable, 
        b: Variable, 
        c: Variable
    ) -> Result<(), PlonkError>;

    fn add(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>;

    fn sub_gate(
        &mut self, 
        a: Variable, 
        b: Variable, 
        c: Variable
    ) -> Result<(), PlonkError>;

    fn sub(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>;

    fn mul_gate(
        &mut self, 
        a: Variable, 
        b: Variable, 
        c: Variable
    ) -> Result<(), PlonkError>;

    fn mul(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>;

    fn bool_gate(&mut self, a: Variable) -> Result<(), PlonkError>;

    fn equal_gate(&mut self, a: Variable, b: Variable) -> Result<(), PlonkError>;

    fn pad_gate(&mut self, n: usize);

    fn support_lookup(&self) -> bool;

    fn create_bool_variable(
        &mut self, 
        val: bool
    ) -> Result<Variable, PlonkError> { ... }

}

An interface for Plonk constraint systems.

Required methods

fn num_gates(&self) -> usize

The number of constraints.

fn num_vars(&self) -> usize

The number of variables.

fn num_inputs(&self) -> usize

The number of public input variables.

fn num_wire_types(&self) -> usize

The number of wire types of the circuit. E.g., UltraPlonk has 4 different types of input wires, 1 type of output wires, and 1 type of lookup wires.

fn public_input(&self) -> Result<Vec<F>, PlonkError>

The list of public input values.

fn check_circuit_satisfiability(
&self,
pub_input: &[F]
) -> Result<(), PlonkError>

Check circuit satisfiability against a public input.

fn create_constant_variable(&mut self, val: F) -> Result<Variable, PlonkError>

Add a constant variable to the circuit; return the index of the variable.

fn create_variable(&mut self, val: F) -> Result<Variable, PlonkError>

Add a variable to the circuit; return the index of the variable.

fn create_public_variable(&mut self, val: F) -> Result<Variable, PlonkError>

Add a public input variable; return the index of the variable.

fn set_variable_public(&mut self, var: Variable) -> Result<(), PlonkError>

Set a variable to a public variable

fn zero(&self) -> Variable

Return a default variable with value zero.

fn one(&self) -> Variable

Return a default variable with value one.

fn witness(&self, idx: Variable) -> Result<F, PlonkError>

Return the witness value of variable idx. Return error if the input variable is invalid.

fn constant_gate(
&mut self,
var: Variable,
constant: F
) -> Result<(), PlonkError>

Common gates that should be implemented in any constraint systems.

Constrain a variable to a constant. Return error if var is an invalid variable.

fn add_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable
) -> Result<(), PlonkError>

Constrain variable c to the addition of a and b. Return error if the input variables are invalid.

fn add(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>

Obtain a variable representing an addition. Return the index of the variable. Return error if the input variables are invalid.

fn sub_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable
) -> Result<(), PlonkError>

Constrain variable c to the subtraction of a and b. Return error if the input variables are invalid.

fn sub(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>

Obtain a variable representing a subtraction. Return the index of the variable. Return error if the input variables are invalid.

fn mul_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable
) -> Result<(), PlonkError>

Constrain variable c to the multiplication of a and b. Return error if the input variables are invalid.

fn mul(&mut self, a: Variable, b: Variable) -> Result<Variable, PlonkError>

Obtain a variable representing a multiplication. Return the index of the variable. Return error if the input variables are invalid.

fn bool_gate(&mut self, a: Variable) -> Result<(), PlonkError>

Constrain a variable to a bool. Return error if the input is invalid.

fn equal_gate(&mut self, a: Variable, b: Variable) -> Result<(), PlonkError>

Constrain two variables to have the same value. Return error if the input variables are invalid.

fn pad_gate(&mut self, n: usize)

Pad the circuit with n dummy gates

fn support_lookup(&self) -> bool

Plookup-related methods. Return true if the circuit support lookup gates.

Provided methods

fn create_bool_variable(&mut self, val: bool) -> Result<Variable, PlonkError>

Add a bool variable to the circuit; return the index of the variable.