jf_plonk::errors

Enum PlonkError

pub enum PlonkError {

    IndexTooLarge,
    DomainCreationError,
    IndexError,
    DivisionError,
    SnarkError(SnarkError),
    PCSError(PCSError),
    CircuitError(CircuitError),
    IoError(Error),
    SerializationError(SerializationError),
    WrongProof,
    RescueError(RescueError),
    InvalidParameters(String),
    NonNativeFieldOverflow,
    IteratorOutOfRange,
    PublicInputsDoNotMatch,

}

A enum specifying the possible failure modes of the Plonk.

Variants

IndexTooLarge

The index is too large for the universal public parameters

DomainCreationError

Failed to create domain

IndexError

Failed to get array value by index

DivisionError

Divided by zero field element

SnarkError(SnarkError)

An error in the Plonk SNARK logic: {0}

PCSError(PCSError)

An error in the underlying polynomial commitment: {0}

CircuitError(CircuitError)

An error in the Plonk circuit: {0}

IoError(Error)

An error during IO: {0}

SerializationError(SerializationError)

An error during (de)serialization

WrongProof

Plonk proof verification failed due to wrong proof

RescueError(RescueError)

Rescue Error

InvalidParameters(String)

Invalid parameters

NonNativeFieldOverflow

Non-native field overflow

IteratorOutOfRange

Iterator out of range

PublicInputsDoNotMatch

Public inputs for partial verification circuit do not match