Module jf_plonk::proof_system::structs

Data structures used in Plonk proof systems

Structs

BatchProof

An aggregated SNARK proof that batchly proving multiple instances.

PlookupEvaluations

A struct that stores the polynomial evaluations in a Plookup argument proof.

PlookupProof

A Plookup argument proof.

PlookupProvingKey

Preprocessed prover parameters used to compute Plookup proofs for a certain circuit.

PlookupVerifyingKey

Preprocessed verifier parameters used to verify Plookup proofs for a certain circuit.

Proof

A Plonk SNARK proof.

ProofEvaluations

A struct that stores the polynomial evaluations in a Plonk proof.

ProvingKey

Preprocessed prover parameters used to compute Plonk proofs for a certain circuit.

VerifyingKey

Preprocessed verifier parameters used to verify Plonk proofs for a certain circuit.

Type Aliases

CommitKey

Commitment key

OpenKey

Key for verifying PCS opening proof.

UniversalSrs

Universal StructuredReferenceString