jf_plonk::proof_system::structs

Struct VerifyingKey

pub struct VerifyingKey<E: Pairing> {
    pub domain_size: usize,
    pub num_inputs: usize,
    pub sigma_comms: Vec<Commitment<E>>,
    pub selector_comms: Vec<Commitment<E>>,
    pub k: Vec<E::ScalarField>,
    pub open_key: OpenKey<E>,
    pub is_merged: bool,
    pub plookup_vk: Option<PlookupVerifyingKey<E>>,
}

Preprocessed verifier parameters used to verify Plonk proofs for a certain circuit.

Fields
domain_size: usize

The size of the evaluation domain. Should be a power of two.

num_inputs: usize

The number of public inputs.

sigma_comms: Vec<Commitment<E>>

The permutation polynomial commitments. The commitments are not hiding.

selector_comms: Vec<Commitment<E>>

The selector polynomial commitments. The commitments are not hiding.

k: Vec<E::ScalarField>

The constants K0, โฆ, K_num_wire_types that ensure wire subsets are disjoint.

open_key: OpenKey<E>

KZG PCS opening key.

is_merged: bool

A flag indicating whether the key is a merged key.

plookup_vk: Option<PlookupVerifyingKey<E>>

Plookup verifying key, None if not support lookup.

Implementations

impl<E, F, P> VerifyingKey<E>
where E: Pairing<BaseField = F, G1Affine = Affine<P>>, F: SWToTEConParam, P: SWCurveConfig<BaseField = F>,

pub fn convert_te_coordinates_to_scalars(&self) -> Vec<F>

Convert the group elements to a list of scalars that represent the Twisted Edwards coordinates.

impl<E: Pairing> VerifyingKey<E>

pub fn dummy(num_inputs: usize, domain_size: usize) -> Self

Create a dummy TurboPlonk verification key for a circuit with num_inputs public inputs and domain size domain_size.