Trait jf_plonk::proof_system::Snark

pub trait Snark<E: PairingEngine> {
    type Proof: Clone;
    type ProvingKey: Clone;
    type VerifyingKey: Clone;
    fn prove<C, R, T>(
        prng: &mut R, 
        circuit: &C, 
        prove_key: &Self::ProvingKey, 
        extra_transcript_init_msg: Option<Vec<u8>>
    ) -> Result<Self::Proof, PlonkError>
    where
        C: Arithmetization<E::Fr>,
        R: CryptoRng + RngCore,
        T: PlonkTranscript<E::Fq>;

    fn verify<T: PlonkTranscript<E::Fq>>(
        verify_key: &Self::VerifyingKey, 
        public_input: &[E::Fr], 
        proof: &Self::Proof, 
        extra_transcript_init_msg: Option<Vec<u8>>
    ) -> Result<(), PlonkError>;
}

An interface for SNARKs.

Associated Types

type Proof: Clone

The SNARK proof computed by the prover.

type ProvingKey: Clone

The parameters required by the prover to compute a proof for a specific circuit.

type VerifyingKey: Clone

The parameters required by the verifier to validate a proof for a specific circuit.

Required methods

fn prove<C, R, T>(
prng: &mut R,
circuit: &C,
prove_key: &Self::ProvingKey,
extra_transcript_init_msg: Option<Vec<u8>>
) -> Result<Self::Proof, PlonkError> where
C: Arithmetization<E::Fr>,
R: CryptoRng + RngCore,
T: PlonkTranscript<E::Fq>,

Compute a SNARK proof of a circuit circuit, using the corresponding proving key prove_key. The witness used to generate the proof can be obtained from circuit.

extra_transcript_init_msg is the optional message to be appended to the transcript during its initialization before obtaining any challenges. This field allows application-specific data bound to the resulting proof without any check on the data. It does not incur any additional cost in proof size or prove time.

fn verify<T: PlonkTranscript<E::Fq>>(
verify_key: &Self::VerifyingKey,
public_input: &[E::Fr],
proof: &Self::Proof,
extra_transcript_init_msg: Option<Vec<u8>>
) -> Result<(), PlonkError>

Verify a SNARK proof proof of the circuit circuit, with respect to the public input pub_input.

extra_transcript_init_msg: refer to documentation of prove