Trait jf_plonk::circuit::customized::rescue::RescueNonNativeGadget

pub trait RescueNonNativeGadget<F: PrimeField> {
    fn rescue_permutation<T: RescueParameter>(
        &mut self, 
        input_var: RescueNonNativeStateVar<F>
    ) -> Result<RescueNonNativeStateVar<F>, PlonkError>;
    fn prp<T: RescueParameter>(
        &mut self, 
        key_var: &RescueNonNativeStateVar<F>, 
        input_var: &RescueNonNativeStateVar<F>
    ) -> Result<RescueNonNativeStateVar<F>, PlonkError>;
    fn rescue_sponge_no_padding<T: RescueParameter>(
        &mut self, 
        data_vars: &[FpElemVar<F>], 
        num_output: usize
    ) -> Result<Vec<FpElemVar<F>>, PlonkError>;
    fn rescue_sponge_with_padding<T: RescueParameter>(
        &mut self, 
        data_vars: &[FpElemVar<F>], 
        num_output: usize
    ) -> Result<Vec<FpElemVar<F>>, PlonkError>;
    fn rescue_full_state_keyed_sponge_no_padding<T: RescueParameter>(
        &mut self, 
        key: FpElemVar<F>, 
        data_vars: &[FpElemVar<F>]
    ) -> Result<FpElemVar<F>, PlonkError>;
}

Trait for rescue circuit over non-native field.

Required Methods

fn rescue_permutation<T: RescueParameter>(
&mut self,
input_var: RescueNonNativeStateVar<F>
) -> Result<RescueNonNativeStateVar<F>, PlonkError>

Given an input state st_0 and an output state st_1, ensure that st_1 = rescue_permutation(st_0) where rescue_permutation is the instance of the Rescue permutation defined by its respective constants

input_var - variables corresponding to the input state

returns - variables corresponding to the output state

fn prp<T: RescueParameter>(
&mut self,
key_var: &RescueNonNativeStateVar<F>,
input_var: &RescueNonNativeStateVar<F>
) -> Result<RescueNonNativeStateVar<F>, PlonkError>

Rescue based Pseudo Random Permutation (PRP)

key_var - rescue state variable corresponding to the cipher key

input_var - rescue state variable corresponding to the plaintext

returns - state variable corresponding to the cipher text

fn rescue_sponge_no_padding<T: RescueParameter>(
&mut self,
data_vars: &[FpElemVar<F>],
num_output: usize
) -> Result<Vec<FpElemVar<F>>, PlonkError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables, data_vars.len() should be a positive integer that is a multiple of the sponge rate (i.e. 3)

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_sponge_with_padding<T: RescueParameter>(
&mut self,
data_vars: &[FpElemVar<F>],
num_output: usize
) -> Result<Vec<FpElemVar<F>>, PlonkError>

Sponge-based hashes from Rescue permutations

data_vars - sponge input variables,

num_output - number of output variables

returns - a vector of variables that refers to the sponge hash output

fn rescue_full_state_keyed_sponge_no_padding<T: RescueParameter>(
&mut self,
key: FpElemVar<F>,
data_vars: &[FpElemVar<F>]
) -> Result<FpElemVar<F>, PlonkError>

Full-State-Keyed-Sponge with a single output

key - key variable

input - input variables,

returns a variable that refers to the output