Module jf_plonk::circuit::gates

Implementations of various Gates.

Structs

AdditionGate

An addition gate

BoolGate

A boolean gate, selectors identical to MultiplicationGate, achieve through constraining a * a = a

ConstantAdditionGate

Adding a variable by a constant.

ConstantGate

A constant gate

ConstantMultiplicationGate

A mul constant gate. Multiply the first variable with the constant.

EqualityGate

An equality gate, selectors identical to SubtractionGate, achieve through constraining a - b = 0

FifthRootGate

Gate for checking a value is the fifth root of another

IoGate

An I/O gate for public inputs

MultiplicationGate

A multiplication gate

PaddingGate

A empty gate for circuit padding

SubtractionGate

A subtraction gate

Traits

Gate

Describes a gate with getter for all selectors configuration

GateClone

Clone a Gate.