Module jf_plonk::proof_system::batch_arg

An argument system that proves/verifies multiple instances in a batch.

Structs

BatchArgument

A batching argument.

Instance

A circuit instance that consists of the corresponding proving key/verification key/circuit.

Functions

build_batch_proof_and_vks_for_test

Create num_instances type A/B instance verifying keys and compute the corresponding batch proof. Only used for testing.