Struct jf_plonk::transcript::RescueTranscript

pub struct RescueTranscript<F>where
    F: RescueParameter,{ /* private fields */ }

Transcript with rescue hash function.

It is currently implemented simply as

an append only vector of field elements

a state that is initialized with 0

We keep appending new elements to the transcript vector, and when a challenge is to be generated, we reset the state with the fresh challenge.

state: [F: STATE_SIZE] = hash(state|transcript)

challenge = state[0]

transcript = vec![challenge]