Trait jf_plonk::transcript::PlonkTranscript

pub trait PlonkTranscript<F> {
    fn new(label: &'static [u8]) -> Self;
    fn append_message(
        &mut self,
        label: &'static [u8],
        msg: &[u8]
    ) -> Result<(), PlonkError>;
    fn get_and_append_challenge<E>(
        &mut self,
        label: &'static [u8]
    ) -> Result<E::Fr, PlonkError>
    where
        E: PairingEngine<Fq = F>;

    fn append_vk_and_pub_input<E, P>(
        &mut self,
        vk: &VerifyingKey<E>,
        pub_input: &[E::Fr]
    ) -> Result<(), PlonkError>
    where
        E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
        P: SWParam<BaseField = F>,
    { ... }
    fn append_commitments<E, P>(
        &mut self,
        label: &'static [u8],
        comms: &[Commitment<E>]
    ) -> Result<(), PlonkError>
    where
        E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
        P: SWParam<BaseField = F>,
    { ... }
    fn append_commitment<E, P>(
        &mut self,
        label: &'static [u8],
        comm: &Commitment<E>
    ) -> Result<(), PlonkError>
    where
        E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
        P: SWParam<BaseField = F>,
    { ... }
    fn append_challenge<E>(
        &mut self,
        label: &'static [u8],
        challenge: &E::Fr
    ) -> Result<(), PlonkError>
    where
        E: PairingEngine<Fq = F>,
    { ... }
    fn append_proof_evaluations<E: PairingEngine>(
        &mut self,
        evals: &ProofEvaluations<E::Fr>
    ) -> Result<(), PlonkError> { ... }
    fn append_plookup_evaluations<E: PairingEngine>(
        &mut self,
        evals: &PlookupEvaluations<E::Fr>
    ) -> Result<(), PlonkError> { ... }
}

Defines transcript APIs.

It has an associated type F which defines the native field for the snark circuit.

The transcript can be either a Merlin transcript (instantiated with Sha-3/keccak), or a Rescue transcript (instantiated with Rescue hash), or a Solidity-friendly transcript (instantiated with Keccak256 hash). The second is only used for recursive snarks.

Required Methods

fn new(label: &'static [u8]) -> Self

Create a new plonk transcript.

fn append_message(
&mut self,
label: &'static [u8],
msg: &[u8]
) -> Result<(), PlonkError>

Append the message to the transcript.

fn get_and_append_challenge<E>(
&mut self,
label: &'static [u8]
) -> Result<E::Fr, PlonkError>where
E: PairingEngine<Fq = F>,

Generate the challenge for the current transcript, and then append it to the transcript.

Provided Methods

fn append_vk_and_pub_input<E, P>(
&mut self,
vk: &VerifyingKey<E>,
pub_input: &[E::Fr]
) -> Result<(), PlonkError>where
E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
P: SWParam<BaseField = F>,

Append the verification key and the public input to the transcript.

fn append_commitments<E, P>(
&mut self,
label: &'static [u8],
comms: &[Commitment<E>]
) -> Result<(), PlonkError>where
E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
P: SWParam<BaseField = F>,

Append a slice of commitments to the transcript.

fn append_commitment<E, P>(
&mut self,
label: &'static [u8],
comm: &Commitment<E>
) -> Result<(), PlonkError>where
E: PairingEngine<Fq = F, G1Affine = GroupAffine<P>>,
P: SWParam<BaseField = F>,

Append a single commitment to the transcript.

fn append_challenge<E>(
&mut self,
label: &'static [u8],
challenge: &E::Fr
) -> Result<(), PlonkError>where
E: PairingEngine<Fq = F>,

Append a challenge to the transcript.

fn append_proof_evaluations<E: PairingEngine>(
&mut self,
evals: &ProofEvaluations<E::Fr>
) -> Result<(), PlonkError>

Append a proof evaluation to the transcript.

fn append_plookup_evaluations<E: PairingEngine>(
&mut self,
evals: &PlookupEvaluations<E::Fr>
) -> Result<(), PlonkError>

Append the plookup evaluation to the transcript.